Resolved the architectural issue that the UseCase would return a Dto instead a domain entity
Mapping the domain entities to dtos are the presentation layer´s responsibility.
The application layer´s use cases shall return domain entities and not care how
the data is then given to the request source.

The reasoning behind this is, that then the use case classes are theoretically
reusable on other parts of the codebase without needing to be reworked
to create a different dto object.

Moved the DTO and Automapper profile for mapping the domain entities to DTOs
to the presentation layer too. The mapping is now done in the controllers
instead of the use cases.