Cleaning up the database seeding
Previously the SeedDataLoader was not loading the data
into entities instead of a SeedData object.

The SeedDatabaseCommandHandler needs to use the entities
becase of the accessibility of the SeedData class. The SeedData
needs to be internal as it makes the Infra entities publicly accessible.
The infra entitites must be internal at all cost otherwise they would
leak into the other layers of the clean architecture.

The TestDatabase create methods were simplified as well.