Replaced UnhandlerExceptionFilter with the CinemaTicketBookingExceptionHandler
This is now a middleware that will be able to handle all
the known exceptions in a single place.
This will enable us to remove the repetitive
try-catch blocks in controller actions.

Unfortunately the developer exception page
cannot be used anymore as now we need the
CinemaTicketBookingExceptionHandler to return
proper HTTP responses.