TierEntity: Fixed connection to Pricings
Previously a tier could have only one pricing, which
is wrong. We shall be able to define multiple pricings
for a tier because different screenings shall have different
price for the same tier.

For example a screening of a blockbuster in the evening
on weekends is probably more expensive than a screening of a
documentary on a weekday.

What is unique is the screening tier combination of a pricing.
This is enforced in the PricingEntity.